protocol/process_param_header: refactor to use pair<,>
process_param_header has been refactored to use
std::pair<std::string, std::string>& to store the result instead of
using (char*&, size_t&, char*&, size_t&).
This method is safer and results in less overhead during parameter
extraction.

The return value has changed from bool to int so that the buffer can be
incrementally chopped by the number of bytes read without re-querying
the lengths of .first and .second and performing size field
calculations.
Now, if the extraction was successful, the number of parsed bytes is
returned.
On length error, -1 is returned.