Delete koi8 and cgi examples.
Koi8 is deleted because usage of iconv or other filter would require
a function hook, which is not yet implemented, so that example is
deleted.

The cgi example is deleted because it's too much work to generate
both the standard and non-fcgi versions of the library from the same
code.