manager: refactor
Manager has been refactored into Manager and ManagerT.

Namely, Manager makes use of Request_base's virtuals and stores pointers
to instances. The virtualness of the dtors prevents memory leaks.
Now, it's template-invariant and the implementation can be moved to
src/.

With the implementation of the per-request management moved to src/, the
library-general portions of management are moved here, too, namely
GET_VALUES handling, which has been moved from src/protocol/funcs.cpp.

The constructor for Manager is
Manager(int fd, std::function<Request_base * ()> new_request).
Because
Manager x([](){return new T;})
can get tedious and ugly, ManagerT<T : Request_base> is defined, which
templatizes the 'return new T;'. Type_traits magic is used to restrict
ManagerT's <T> to only instantiate subtypes of Request_base.

Manager also has a unified signal handler. Since there's only one
Manager instance per process, we can use a global map of pid->Manager*
to co-ordinate all the instance pointers.
Now, it uses the result of getpid to find the correct instance for which
to call .terminate() or .stop().
Pid->instance* registration is done in the constructor and
deregistration in the destructor.
To avoid UB, an assertion is done to ensure that the pointer is
non-zero.