examples/timer: minor refactoring and compile fixes
Remove the redudant Header since content_type already returns a header.
While there's no copy overhead due to move semantics, the less code
there is, the better.

Change the stored message allocator to use uchar to resolve type
conversion warnings.

Change the async_wait lambda to bind the this pointer and take a boost
error code as an argument to fit the function definition expected for
boost::asio::deadline_timer::Handler.

Change the scope of the 'return true' to remove the "not all codepaths
return" warning.